+++ frontend/.env
@@ -1,4 +1,4 @@
-VITE_BFF_FLAG=false
+VITE_BFF_FLAG=true
 VITE_BFF_API_URL=http://localhost:8080/
 
 VITE_BACKEND_API_URL=http://localhost:9500/
